Refactor main class and clean warnings

+++ Point.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 #include "Point.h"
 
-Point::Point(double x, double y) {
+Point::Point(float x, float y) {
     this->x = x;
     this->y = y;
     r = std::sqrt(x*x + y*y);
@@ -13,15 +13,15 @@ Point::Point(double x, double y) {
         fi = 0.0;
     }
     fi *= 180./M_PI;
-    if (x < 0)
-        fi = 180. + fi;
+    if (x < 0.)
+        fi = 180.f + fi;
 };
 
-double Point::getX() {
+float Point::getX() {
     return x;
 }
 
-double Point::getY() {
+float Point::getY() {
     return y;
 }
 
@@ -29,6 +29,6 @@ int Point::operator<(Point point) {
     return this->fi < point.fi;
 }
 
-double Point::distance(Point point) {
-    return sqrt(pow(point.getX() - getX(),2.0) + pow(point.getY() - getY(),2.0));
+float Point::distance(Point point) {
+    return (float)sqrt(pow(point.getX() - getX(),2.0) + pow(point.getY() - getY(),2.0));
 }
